backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
google-api-python-client==2.110.0
requests==2.31.0
cryptography==41.0.7